update install script

+++ scripts/prereqs.sh
@@ -13,19 +13,23 @@ if ! grep -q 'Ubuntu' /etc/issue
 fi
 
 ## Update OS
-sudo apt update  > /dev/null 2>&1
-sudo apt upgrade -y  > /dev/null 2>&1
+sudo apt update
+sudo apt upgrade -y
 
 ## Install Prereqs
 sudo apt-get install -y acl apache2-utils apt-transport-https bash-completion build-essential ca-certificates certbot curl fonts-firacode \
 git gnupg haveged htop jq libnss3-tools libfontconfig1 libgtk-3-0 libxi6 libxrender1 libxtst6 locales lsb-release net-tools openssl python3-pip \
-software-properties-common telnet tree unzip zsh zsh-autosuggestions zip > /dev/null 2>&1
+software-properties-common telnet tree unzip zsh zsh-autosuggestions zip
 
 ## Create user
 echo "Creating Minio user..."
-sudo adduser $USER
-sudo usermod -aG sudo $USER
-echo '$ a $USER ALL=(ALL:ALL) NOPASSWD: ALL' | EDITOR="sed -f- -i" visudo
+if id $USER &>/dev/null ; then
+    echo "$USER exists. No need to add user."
+else
+    sudo adduser --disabled-password --gecos "" $USER
+    sudo usermod -aG sudo $USER
+    echo '$ a ${USER} ALL=(ALL:ALL) NOPASSWD: ALL' | EDITOR="sed -f- -i" visudo
+fi
 
 ## Change Shell
 echo "Installing zsh and oh-my-zsh..."
@@ -33,6 +37,7 @@ FILE=$HOME/.oh-my-zsh/oh-my-zsh.sh
 if [ -f "$FILE" ]; then
     echo "$FILE exists. No need to install oh-my-zsh again."
 else
+    su minio
     sudo mkdir -p /home/$USER/misc/dotfiles
     sudo chmod 0777 -R /home/$USER
     git clone https://github.com/AlphaBravoCompany/ab-dotfiles.git /home/$USER/misc/dotfiles/
@@ -40,6 +45,7 @@ else
     cd /home/$USER/misc/dotfiles
     ./install.sh
     cd ~
+    exit
 fi
 
 ## Docker Things ##
@@ -48,7 +54,7 @@ echo "Installing Docker Engine..."
 FILE=/usr/bin/docker
 if [ -f "$FILE" ]; then
     echo "$FILE exists. No need to install Docker again."
-else 
+else
     echo "$FILE does not exist. Installing Docker."
     curl -fsSL https://get.docker.com -o /tmp/get-docker.sh > /dev/null 2>&1
     sudo sh /tmp/get-docker.sh > /dev/null 2>&1
@@ -60,17 +66,12 @@ echo "Installing Docker Compose..."
 curl -SL https://github.com/docker/compose/releases/download/v2.10.1/docker-compose-linux-x86_64 -o /usr/local/bin/docker-compose > /dev/null 2>&1
 sudo chmod +x /usr/local/bin/docker-compose > /dev/null 2>&1
 
-## Create folder structure
-sudo mkdir -m 777 -p /mnt/data
-sudo mkdir -p /home/$USER/minio/misc/registry
-sudo mkdir -p /home/$USER/minio/tools/deploy/docker/docker-compose/single
-sudo mkdir -p /home/$USER/minio/tools/deploy/docker/docker-compose/distributed
-sudo mkdir -p /home/$USER/minio/tools/deploy/native/
-sudo mkdir -p /home/$USER/minio/tools/deploy/mc/
-sudo mkdir -p /home/$USER/minio/tools/deploy/k8s/operator/
-sudo chmod 0777 -R /home/$USER/minio/
-
 ## Install mkcert
-curl -JLO "https://dl.filippo.io/mkcert/latest?for=linux/amd64"
-chmod +x mkcert-v*-linux-amd64
-sudo cp mkcert-v*-linux-amd64 /usr/local/bin/mkcert
+FILE=/usr/local/bin/mkcert
+if [ -f "$FILE" ]; then
+    echo "$FILE exists. No need to install mkcert again."
+else
+    curl -JLO "https://dl.filippo.io/mkcert/latest?for=linux/amd64"
+    chmod +x mkcert-v*-linux-amd64
+    sudo cp mkcert-v*-linux-amd64 /usr/local/bin/mkcert
+fi
